fix html and add css
