+++ purge_tournament/assets/js/index.js
@@ -1,13 +1,3 @@
-import { createTournament } from './tournament.js'
-import { editTournament } from './tournament.js'
-
-
-
-
 window.addEventListener('DOMContentLoaded', function() {
-
-
-    createTournament();
-    editTournament();
 
 });
