+++ purge_tournament/assets/js/tournamentEdit.js
@@ -1,31 +1,3 @@
-function createTournament() {
-    let selects = document.getElementsByClassName("top32");
-
-
-    for (let i = 0; i < selects.length; i++) {
-
-        selects[i].addEventListener('change', function(e) {
-
-            let teamId = e.target.value;
-
-            console.log(teamId);
-
-            fetch(`https://valmontpehautpietri.sites.3wa.io/res03-projet-final/purge_tournament/admin/tournaments/create/add-team/${teamId}`)
-                .then(response => response.json())
-                .then(data => {
-                        console.log(`La team ${data.name} est ajoutée`);
-                    }
-
-                );
-
-            e.target.setAttribute('disabled', '');
-            //e.target.option.setAttribute('disabled', '');
-
-        });
-
-    }
-}
-
 function editTournament() {
 
     let tournament = document.getElementById('nbTournamentSpan');
@@ -158,5 +130,8 @@ function editTournament() {
     })
 }
 
-export { createTournament };
-export { editTournament };
+window.addEventListener('DOMContentLoaded', function() {
+
+    editTournament();
+
+});
