+++ purge_tournament/assets/styles/sass/forms/_login.scss
@@ -1,0 +1,74 @@
+// Login
+#login {
+    font-family: 'Montserrat', sans-serif;
+    color: $primary-black;
+    padding: 20px;
+    height: 100%;
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    justify-content: center;
+
+    h3 {
+        margin-top: 0;
+        margin-bottom: 10px;
+        font-size: 1.5rem;
+        text-align: center;
+        font-weight: bold;
+        color: $primary-red;
+    }
+
+    form {
+        display: flex;
+        flex-direction: column;
+        align-items: center;
+        background-color: $primary-yellow;
+        padding: 10vh 10vw;
+
+    }
+
+    fieldset {
+        border: none;
+        margin: 10px 0;
+        width: 100%;
+
+        label {
+            display: block;
+            font-size: 1.2rem;
+            margin-bottom: 5px;
+        }
+
+        input {
+            width: 100%;
+            padding: 5px;
+            border-radius: 5px;
+            border: 1px solid $primary-grey;
+            font-size: 1.2rem;
+
+            &::placeholder {
+                color: $primary-grey;
+            }
+        }
+
+        input[type="submit"] {
+            background-color: $primary-black;
+            color: $primary-white;
+            border: none;
+            padding: 10px 20px;
+            border-radius: 5px;
+            font-size: 1.2rem;
+            cursor: pointer;
+            margin-top: 10px;
+        }
+    }
+
+    @media screen and (min-width: 768px) {
+        display: flex;
+        flex-direction: row;
+        justify-content: center;
+
+        form {
+            margin-left: 20px;
+        }
+    }
+}
