+++ purge_tournament/assets/styles/sass/partials/_footer.scss
@@ -1,0 +1,47 @@
+.main-footer {
+    background-color: $primary-black;
+    color: $primary-white;
+    text-align: center;
+    padding: 20px;
+
+    .footer-content {
+        display: flex;
+        flex-direction: column;
+        justify-content: center;
+        align-items: center;
+
+        p {
+            margin: 0 0 10px;
+        }
+
+        .social-links {
+            display: flex;
+            flex-direction: row;
+            justify-content: center;
+            align-items: center;
+            margin: 0;
+            padding: 0;
+
+            li {
+                list-style: none;
+                margin: 0 10px;
+
+                a {
+                    color: $primary-white;
+                    font-size: 20px;
+                }
+            }
+        }
+    }
+
+    @media (min-width: 768px) {
+        .footer-content {
+            flex-direction: row;
+
+            p {
+                margin: 0;
+                margin-right: 20px;
+            }
+        }
+    }
+}
