+++ purge_tournament/assets/styles/sass/tournaments/_tournaments.scss
@@ -1,0 +1,131 @@
+#admin-teams-list {
+    max-width: 100%;
+    margin: 0 auto;
+    padding: 2rem;
+    background-color: $primary-white;
+
+    h2 {
+        font-size: 2rem;
+        color: $primary-black;
+        margin-bottom: 2rem;
+        font-weight: bold;
+    }
+
+    table {
+        width: 100%;
+        border-collapse: collapse;
+
+        thead {
+            tr {
+                th {
+                    color: $primary-black;
+                    text-align: left;
+                    padding: 1rem;
+                    border-bottom: 1px solid $primary-grey;
+                }
+            }
+        }
+
+        tbody {
+            tr {
+                td {
+                    padding: 1rem;
+                    border-bottom: 1px solid $primary-grey;
+                    color: $primary-black;
+
+                    &:first-child {
+                        font-weight: bold;
+                    }
+
+                    a {
+                        color: $primary-black;
+                        text-decoration: none;
+
+                        &:hover {
+                            color: $primary-red;
+                        }
+                    }
+
+                    button {
+                        background-color: $primary-yellow;
+                        border: none;
+                        color: $primary-black;
+                        padding: 0.5rem 1rem;
+                        border-radius: 5px;
+                        cursor: pointer;
+
+                        &:hover {
+                            background-color: $primary-red;
+                            color: $primary-white;
+                        }
+                    }
+                }
+            }
+        }
+    }
+
+    button {
+        margin-top: 2rem;
+        display: block;
+        width: 100%;
+        max-width: 20rem;
+        padding: 1rem;
+        background-color: $primary-yellow;
+        color: $primary-black;
+        border: none;
+        border-radius: 5px;
+        cursor: pointer;
+
+        &:hover {
+            background-color: $primary-red;
+            color: $primary-white;
+        }
+
+        a {
+            text-decoration: none;
+            color: $primary-black;
+        }
+    }
+
+    @media (max-width: 768px) {
+        padding: 1rem;
+
+        h2 {
+            font-size: 1.5rem;
+        }
+
+        table {
+            thead {
+                tr {
+                    th {
+                        font-size: 0.8rem;
+                        padding: 0.5rem;
+                    }
+                }
+            }
+
+            tbody {
+                tr {
+                    td {
+                        font-size: 0.8rem;
+                        padding: 0.5rem;
+
+                        a {
+                            font-size: 0.8rem;
+                        }
+
+                        button {
+                            font-size: 0.8rem;
+                            padding: 0.3rem 0.6rem;
+                        }
+                    }
+                }
+            }
+        }
+
+        button {
+            font-size: 1rem;
+            padding: 0.8rem;
+        }
+    }
+}
