+++ purge_tournament/templates/private/partials/_footer.phtml
@@ -1,0 +1,11 @@
+<footer class="main-footer">
+  <div class="footer-content">
+    <p>&copy; <?= date('Y'); ?> My Website. All Rights Reserved.</p>
+    <ul class="social-links">
+      <li><a href="#"><i class="fa fa-twitter"></i></a></li>
+      <li><a href="#"><i class="fa fa-facebook"></i></a></li>
+      <li><a href="#"><i class="fa fa-instagram"></i></a></li>
+      <li><a href="#"><i class="fa fa-twitch"></i></a></li>
+    </ul>
+  </div>
+</footer>
