+++ purge_tournament/templates/private/partials/_header.phtml
@@ -1,3 +1,9 @@
-<a href="/res03-projet-final/purge_tournament/logout">Se déconnecter</a>
-<a href="/res03-projet-final/purge_tournament/admin/teams">Teams</a>
-<a href="/res03-projet-final/purge_tournament/admin/tournaments">Tournaments</a>
+<header id="admin-header">
+  <nav>
+    <ul>
+      <li><a href="/res03-projet-final/purge_tournament/logout">Se déconnecter</a></li>
+      <li><a href="/res03-projet-final/purge_tournament/admin/teams">Teams</a></li>
+      <li><a href="/res03-projet-final/purge_tournament/admin/tournaments">Tournaments</a></li>
+    </ul>
+  </nav>
+</header>
