+++ purge_tournament/templates/private/private_layout.phtml
@@ -31,8 +31,6 @@
 <?php
     require 'partials/_footer.phtml'; // Je charge le footer dans le dossier partials.
 ?>
-
-<script type="module" src="/res03-projet-final/purge_tournament/assets/js/index.js"></script>
 <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0-alpha1/dist/js/bootstrap.bundle.min.js" integrity="sha384-w76AqPfDkMBDXo30jS1Sgez6pr3x5MlQ1ZAGC+nuZB+EYdgRZgiwxhTBTkF7CXvN" crossorigin="anonymous"></script>
 </body>
 </html>
