+++ purge_tournament/templates/private/tournaments/create.phtml
@@ -37,5 +37,6 @@
     </section>
     <input type="submit" value="Create" name="create" id="tournament-create-btn"/>
     <input type="hidden" name="formName" value="tournament-create"/>
-    <button id="reset-tournament">Reset</button>
+    <button id="reset-tournament">reset</button>
 </form>
+<script type="module" src="/res03-projet-final/purge_tournament/assets/js/tournamentCreate.js"></script>
