+++ purge_tournament/templates/private/tournaments/edit/edit.phtml
@@ -108,5 +108,7 @@
     <input type="hidden" name="formName" value="tournament-edit-2" idTournoi="<?= $data['tournament']['id']?>" id="tournamentId"/>
 </form>
 <button id="reset-tournament">Reset</button>
+<script type="module" src="/res03-projet-final/purge_tournament/assets/js/tournamentEdit.js"></script>
+
