+++ purge_tournament/templates/public/contact/_contact-form.phtml
@@ -1,22 +1,22 @@
 <section id="contact-form">
-    <h1 class="sr-only">Contact Form</h1>
+    <h1 class="sr-only">Formulaire de contact</h1>
     <form action="/res03-projet-final/purge_tournament/contact" method="post">
-        <h2>Contact Us</h2>
+        <h2>Contactez nous !</h2>
         <fieldset>
-            <label for="contactName">Full Name</label>
-            <input type="text" name="contactName"  placeholder="Type your full Name" />
+            <label for="contactName">Nom & prénom</label>
+            <input type="text" name="contactName"  placeholder="Votre nom & prénom" />
         </fieldset>
         <fieldset>
             <label for="contactEmail">Email</label>
-            <input type="email" name="contactEmail"  placeholder="Type your email" />
+            <input type="email" name="contactEmail"  placeholder="Votre email" />
         </fieldset>
         <fieldset>
             <label for="contactMessage">Message</label>
-            <textarea name="contactMessage"rows="5" cols="33">Your message ...
+            <textarea name="contactMessage"rows="5" cols="33" placeholder="Your message ...">
             </textarea>
         </fieldset>
         <fieldset>
-            <input type="submit" name="send" value="Send" id="contact-btn" />
+            <input type="submit" name="Envoyer" value="Envoyer" id="contact-btn" />
         </fieldset>
         <input type="hidden" name="formName" value="contact"/>
     </form>
