+++ purge_tournament/templates/public/contact/contact.phtml
@@ -1,5 +1,3 @@
 <?php
 
-var_dump($data);
-
 require '_contact-form.phtml';
