+++ purge_tournament/templates/public/login/_login-form.phtml
@@ -1,14 +1,13 @@
 <section id="login">
-    <h1 class="sr-only">Login</h1>
     <form action="/res03-projet-final/purge_tournament/admin" method="post">
-        <h2>Login</h2>
+        <h3>Connexion</h3>
         <fieldset>
             <label for="email">Email</label>
-            <input type="email" name="email"  placeholder="Type your email" />
+            <input type="email" name="email"  placeholder="Email administrateur" />
         </fieldset>
         <fieldset>
-            <label for="password">Password</label>
-            <input type="password" name="password"  placeholder="Type your password" />
+            <label for="password">Mot de passe</label>
+            <input type="password" name="password"  placeholder="Mot de passe" />
         </fieldset>
         <fieldset>
             <input type="submit" name="submit" value="Se connecter" id="login-button" />
