+++ purge_tournament/templates/public/login/login.phtml
@@ -1,7 +1,5 @@
-<section id="login">
+<h2 class="sr-only">Login</h2>
 <?php
-
-var_dump($data);
 
 require '_login-form.phtml';
