+++ purge_tournament/templates/public/partials/_footer.phtml
@@ -1,1 +1,11 @@
-<footer>Footer</footer>
+<footer class="main-footer">
+  <div class="footer-content">
+    <p>&copy; <?= date('Y'); ?> Valmonzo. All Rights Reserved.</p>
+    <ul class="social-links">
+      <li><a href="#"><i class="fa-brands fa-twitter"></i></a></li>
+      <li><a href="#"><i class="fa-brands fa-facebook"></i></a></li>
+      <li><a href="#"><i class="fa-brands fa-instagram"></i></a></li>
+      <li><i class="fa-brands fa-twitch"></i></a></li>
+    </ul>
+  </div>
+</footer>
